requirements.txt
twilio==8.10.0
pandas==2.1.4
python-dotenv==1.0.0
tkinter
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
werkzeug==3.0.1
mutagen==1.47.0
pydub==0.25.1

# Email provider dependencies
sendgrid==6.11.0
mailgun==0.1.1
boto3==1.34.0  # For Amazon SES
email-validator==2.1.0
